+++ CHANGELOG.md
@@ -3,3 +3,7 @@ Changelog is normally kept [here](https://github.com/Cervi-Robotics/interceptor/
 This is only for new-version changes:
 
 # v1.2
+
+* added `notify` and `unnotify` command
+* added the `notify_about_actions` configuration
+